requirements.txt
python-telegram-bot[jobqueue,webhooks]==21.7
watchdog
yt-dlp
edge-tts